Adds the draw and graphicDecode functions, adds snake_head struct
Declares some intial variables in snake.c, declares and implements the draw function, which in combination with the graphicDecode function display the current state of the game to the LCD display.
The snake_head.h header file only contains a simple struct, snake_head, which holds the x and y coordinates of the head of the snake.