Rewrites movement fucntion, adds fruit_get() function
The movement fucntion in movement.c now moves the tail if the snake did not eat fruit. It also now returns a boolean representing whether a fruit has been eaten or not.
The snake.c received a function which now triggers after the snake eats a fruit. It increments the score (shown on the led line below the display) and flashes a green ligh on the RGB diodes.
If max score is reached (the led line fills up, ie the score == 32), a short celebration occurs in which the led line flashes, and the RGB diodes turn yellow. The score then resets.