Implements variable snake width, moves knob parsing to separate module, adds once move lockout, adjusts game speed
After initial testing, a snake width of one pixel was far too small for practial usage. Therefore, snake size can now be adujsted in the global SNAKE_SIZE variable.
This affects how many pixels wide the snake will be, as well as how big the fruit will be.
The default size is 5 pixels, and only odd pixel sizes have been tested, even widths might not work as intended.

Knob parsing was moved to a separate module, as it was proving too complicated for one function in the game loop.
It now also correctly checks for char overflow on the knob value.

The tick (one run of the game loop) after a player has inputted a direction, the snake will always move the direction it was heading before, effectivelly adding a one tick lockout to player movement.
This was done because at higher speeds (and sometimes normal speeds) the snake would just circle into itself, ending the game, even though the inputted turn was not that significant.
This change should make moving the snake more intuitive.

Also adjusted tick speed. The original clock delay was too large, so it overflowed into the shortest one possible. The delay was reduces, and thanks to it no longer overflowing, the game got a lot slower.
As compensation, the game now speeds up much more when picking up fruit.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch two-players
# Your branch is up to date with 'origin/two-players'.
#
# Changes to be committed:
#	new file:   .vscode/configurationCache.log
#	new file:   .vscode/dryrun.log
#	new file:   .vscode/settings.json
#	new file:   .vscode/targets.log
#	modified:   Makefile
#	modified:   fruit_generation.c
#	modified:   fruit_generation.h
#	new file:   knob_parsing.c
#	new file:   knob_parsing.h
#	modified:   movement.c
#	modified:   movement.h
#	modified:   mzapo_parlcd.c
#	modified:   snake
#	modified:   snake.c
#