Removes pause button, adjusts snake_death_animation to work properly on self destruct
Removed the pause function from the green knob press as it was making entering the 1x speed mode (also green knob) difficult, and it did not serve much use.

Reworked how snake_death_animation works, not triggers once outside of loop. This is because after a collision with wall or snake, the snake_head of the snake now moves back a position.
This allows the snake animation to work properly even when the snake self destructs (makes itself into a loop).